move test util functions
