+++ src/storage/txn/actions/mod.rs
@@ -5,6 +5,7 @@
 //! eg. [`MvccTxn::load_lock`], [`MvccTxn::put_write`], which are methods on [`MvccTxn`],
 //! for archiving a certain target
 
+pub mod acquire_pessimistic_lock;
 pub mod commit;
 pub mod pessimistic_prewrite;
 pub mod prewrite;
