+++ src/storage/txn/commands/acquire_pessimistic_lock.rs
@@ -8,7 +8,7 @@ use crate::storage::mvcc::{Error as MvccError, ErrorInner as MvccErrorInner, Mvc
 use crate::storage::txn::commands::{
     Command, CommandExt, TypedCommand, WriteCommand, WriteContext, WriteResult,
 };
-use crate::storage::txn::{Error, ErrorInner, Result};
+use crate::storage::txn::{acquire_pessimistic_lock, Error, ErrorInner, Result};
 use crate::storage::{
     Error as StorageError, ErrorInner as StorageErrorInner, PessimisticLockRes, ProcessResult,
     Result as StorageResult, Snapshot,
@@ -84,7 +84,8 @@ impl<S: Snapshot, L: LockManager> WriteCommand<S, L> for AcquirePessimisticLock 
             Ok(PessimisticLockRes::Empty)
         };
         for (k, should_not_exist) in keys {
-            match txn.acquire_pessimistic_lock(
+            match acquire_pessimistic_lock(
+                &mut txn,
                 k,
                 &self.primary,
                 should_not_exist,
